requirements.txt
Pillow>=10.0.0
PyMuPDF>=1.23.0
pyinstaller>=6.0.0
